Combine both conversation fetchers
